Admins cannot delete admin user
Added functionality so that the admin cannot delete the first/main admin user

+++ includes/deleteadm_ngouser.inc.php
@@ -9,16 +9,19 @@ if ($_SERVER["REQUEST_METHOD"] == "POST"){
         require_once 'config_session.inc.php';
         require_once 'dbh.inc.php';
 
-        $query = "DELETE FROM glngousers WHERE id=$ngo_id";
-        $pdo->exec($query);
-
-        header("Location: ../Private/p_adminpanel/adminngopanel.php");
-
-        //close connections
-        $pdo = null;
-        $stmt = null;
-        die();
-
+        if($ngo_id != 1){
+            $query = "DELETE FROM glngousers WHERE id=$ngo_id";
+            $pdo->exec($query);
+    
+            header("Location: ../Private/p_adminpanel/adminngopanel.php");
+    
+            //close connections
+            $pdo = null;
+            $stmt = null;
+            die();
+        } else{
+            header("Location: ../Private/p_adminpanel/adminngopanel.php");
+        }
 
     } catch (PDOException $e) {
         die("Query failed (Delete user page): " . $e->getMessage());
